backend/requirements.txt
celery==5.2.3
django-cors-headers==3.11.0
Django==4.0.1
djangorestframework-simplejwt==5.0.0
djangorestframework==3.13.1
environs==9.4.0
psycopg2==2.9.1
redis==4.1.2